Remove broken file asset urls
There are some urls burried in a random scss file. I commented out the
bad urls and replaced them with url("") which doesn't give an error and
the sample website doesn't look any different. If it becomes a problem
we can revert it later but we need it right now in order to run the
website. >30 math depreciations left!